feat(online-evals): first-run empty state for the project evaluators tab
The gallery rendered in place of `<tbody>`, so the Name/Target/Filter/
Sampling/Enabled/Actions headers sat above it and a project with no
evaluators read as a broken table rather than a first-run experience. Its
one line of copy, "No evaluators are set up for this project", was buried
under the cards in italics.

The gallery itself was right, so it is kept as it was: two columns, LLM on
the left and code on the right, each led by a "Create new ..." card and
followed by the instance's existing evaluators to copy or attach. What
changes is where it sits and what leads it. It now replaces the table
outright, under an EmptyState -- the evaluator graphic, a title, and a
sentence on what an evaluator does -- which is the dataset evaluators
empty state's shape, so the two surfaces read as the same thing.

The copy covers all three targets rather than only spans, since an
evaluator scores spans, traces, or sessions.

The gallery resolves behind a skeleton of the same two rows of cards, so
the header above it does not move, and an error there is a single line of
text rather than a torn layout.

A search that matches nothing keeps the table, since the columns still
describe what is being searched.
